requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.0
whitenoise==5.3.0
